backend/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
drf-nested-routers==0.92.5
PyJWT==2.0.1
pytz==2020.5
sqlparse==0.4.1
